+++ complex/lib/core.php
@@ -2,6 +2,8 @@
 /**
  * WebMan WordPress Theme Framework
  *
+ * Textdomain used in the framework: wmtf_domain
+ *
  * @copyright  2015 WebMan - Oliver Juhas
  * @license    GPL-2.0+, http://www.gnu.org/licenses/gpl-2.0.html
  *