+++ complex/lib/inc/admin.php
@@ -29,7 +29,7 @@
 	//Theme Updater
 
 		if ( apply_filters( 'wmhook_enable_update_notifier', false ) ) {
-			locate_template( WM_LIBRARY_DIR . 'inc/update-notifier/update-notifier.php', true );
+			locate_template( WM_LIBRARY_DIR . 'inc/update-notifier.php', true );
 		}
 
 	//Admin class