+++ complex/lib/js/customizer.js
@@ -18,12 +18,16 @@
 	 * Custom radio select
 	 *
 	 * @since   3.1
-	 * @version 3.1
+	 * @version 5.0
 	 */
 
 	jQuery( '.custom-radio-container' ).on( 'change', 'input', function() {
 
-			jQuery( this ).parent().addClass( 'active' ).siblings().removeClass( 'active' );
+			jQuery( this )
+				.parent()
+					.addClass( 'active' )
+					.siblings()
+					.removeClass( 'active' );
 
 		} );