+++ simple/customizer/controls/class-Customizer_Hidden.php
@@ -1,0 +1,33 @@
+<?php
+/**
+ * Customizer custom controls
+ *
+ * Customizer hidden input field.
+ *
+ * @package    Receptar
+ * @copyright  2015 WebMan - Oliver Juhas
+ *
+ * @since    1.0
+ * @version  1.0
+ */
+
+
+
+/**
+ * Hidden
+ */
+class Receptar_Customizer_Hidden extends WP_Customize_Control {
+
+	public $type = 'hidden';
+
+	public function render_content() {
+		?>
+
+		<textarea <?php $this->link(); ?>><?php echo esc_textarea( $this->value() ); ?></textarea>
+
+		<?php
+	}
+
+} // /Receptar_Customizer_Hidden
+
+?>