+++ complex/lib/changelog.md
@@ -1,12 +1,20 @@
-# Changelog
+# WebMan WordPress Theme Framework Changelog
 
 ## 5.0
 
 * UPDATED: Code organized into PHP classes
+* UPDATED: Replaced pluggable functions functionality with "pre" hook
+* UPDATED: Removed obsolete constants and functions
+* UPDATED: Simplified code
+* UPDATED: Renamed textdomain to `wmtf_domain`
+* UPDATED: WordPress 4.1+ compatible only
+* UPDATED: Scripts: TGM Plugin Activation 2.4.2
+* UPDATED: Removed customizer skin creator/loader in favour for [Customizer Export/Import plugin](https://wordpress.org/plugins/customizer-export-import/)
+* UPDATED: Renamed hooks
 
 #### Files changed:
 
-
+	*.* (All)
 
 
 ## 4.0.4